+++ apps/webapp/src/app/pages/memes/memes.component.ts
@@ -1,6 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { MemesService } from './memes.service';
 import { IImageMetadata } from '@portfolio/api-interfaces';
+import { AddMemeDialogComponent } from './add-meme-dialog/add-meme-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
+import { MemeFormData } from './model/meme.model';
+import { BinaryConfirmationComponent } from '../../modules/custom-dialog/binary-confirmation/binary-confirmation.component';
 
 @Component({
   selector: 'portfolio-memes',
@@ -8,32 +12,48 @@ import { IImageMetadata } from '@portfolio/api-interfaces';
   styleUrls: ['./memes.component.scss'],
 })
 export class MemesComponent implements OnInit {
-  imageObj: File;
-  description = 'sample caption';
-  title = 'sample title';
-
   ngOnInit(): void {}
 
-  constructor(private readonly memesService: MemesService) {}
+  constructor(private readonly memesService: MemesService, public dialog: MatDialog) {}
 
-  onImagePicked(event: Event): void {
-    const FILE = (event.target as HTMLInputElement).files[0];
-    this.imageObj = FILE;
+  public onAddNewMeme() {
+    const dialogRef = this.dialog.open(AddMemeDialogComponent);
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.onImageUpload(result);
+      }
+    });
   }
 
-  get memes$() {
+  public get memes$() {
     return this.memesService.getMemes$();
   }
 
-  onDeleteMeme(meme: IImageMetadata) {
-    this.memesService.deleteMeme(meme);
+  public onDeleteMeme(meme: IImageMetadata) {
+    const dialogProp = {
+      title: 'Delete Meme',
+      messages: ['Are you sure you want to delete ' + meme.title],
+      okayLabel: 'Okay',
+      noLabel: 'Cancel',
+    };
+
+    const dialogRef = this.dialog.open(BinaryConfirmationComponent, {
+      data: dialogProp,
+    });
+
+    dialogRef.afterClosed().subscribe((isTrue: boolean) => {
+      if (isTrue) {
+        this.memesService.deleteMeme(meme);
+      }
+    });
   }
 
-  onImageUpload() {
+  private onImageUpload(formData: MemeFormData) {
     const imageForm = new FormData();
-    imageForm.append('image', this.imageObj);
-    imageForm.append('caption', this.description);
-    imageForm.append('title', this.title);
+    imageForm.append('image', formData.image);
+    imageForm.append('caption', formData.description);
+    imageForm.append('title', formData.title);
     this.memesService.imageUpload(imageForm).subscribe((res) => {
       this.memesService.addMeme(res);
     });
